Rename TropoDB memtable files
