+++ implementation/rocksdb/db/zns_impl/persistence/zns_wal_manager.ipp
@@ -9,7 +9,7 @@
 #include "db/write_batch_internal.h"
 #include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/io/szd_port.h"
-#include "db/zns_impl/memtable/zns_memtable.h"
+#include "db/zns_impl/memtable/tropodb_memtable.h"
 #include "db/zns_impl/persistence/zns_committer.h"
 #include "db/zns_impl/persistence/zns_wal.h"
 #include "db/zns_impl/persistence/zns_wal_manager.h"
