+++ implementation/rocksdb/db/zns_impl/table/l0_zns_sstable.h
@@ -4,7 +4,7 @@
 #define L0_ZNS_SSTABLE_H
 
 #include "db/zns_impl/tropodb_config.h"
-#include "db/zns_impl/memtable/zns_memtable.h"
+#include "db/zns_impl/memtable/tropodb_memtable.h"
 #include "db/zns_impl/persistence/zns_committer.h"
 #include "db/zns_impl/table/zns_sstable.h"
 #include "db/zns_impl/table/zns_sstable_builder.h"
