+++ implementation/rocksdb/db/zns_impl/table/ln_zns_sstable.h
@@ -4,7 +4,7 @@
 #define LN_ZNS_SSTABLE_H
 
 #include "db/zns_impl/tropodb_config.h"
-#include "db/zns_impl/memtable/zns_memtable.h"
+#include "db/zns_impl/memtable/tropodb_memtable.h"
 #include "db/zns_impl/table/zns_sstable.h"
 #include "db/zns_impl/table/zns_sstable_builder.h"
 #include "db/zns_impl/table/zns_zonemetadata.h"
