+++ implementation/rocksdb/db/zns_impl/table/zns_sstable_manager.h
@@ -8,7 +8,7 @@
 #include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/utils/tropodb_diagnostics.h"
 #include "db/zns_impl/io/szd_port.h"
-#include "db/zns_impl/memtable/zns_memtable.h"
+#include "db/zns_impl/memtable/tropodb_memtable.h"
 #include "db/zns_impl/ref_counter.h"
 #include "db/zns_impl/table/l0_zns_sstable.h"
 #include "db/zns_impl/table/ln_zns_sstable.h"
