+++ implementation/rocksdb/db/zns_impl/tropodb_impl.h
@@ -28,7 +28,7 @@
 #include "db/zns_impl/index/tropodb_version.h"
 #include "db/zns_impl/index/tropodb_version_set.h"
 #include "db/zns_impl/io/szd_port.h"
-#include "db/zns_impl/memtable/zns_memtable.h"
+#include "db/zns_impl/memtable/tropodb_memtable.h"
 #include "db/zns_impl/persistence/zns_manifest.h"
 #include "db/zns_impl/persistence/zns_wal.h"
 #include "db/zns_impl/persistence/zns_wal_manager.h"
